Refactoring parse/encode/types Also added unproven feature to encoding
