+++ Cargo.toml
@@ -8,6 +8,9 @@ name = "svd-parser"
 repository = "https://github.com/japaric/svd"
 version = "0.6.0"
 
+[features]
+unproven = []
+
 [dependencies]
 either = "1.1.0"
 xmltree = "0.3.2"
