+++ src/svd/modifiedwritevalues.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 use xmltree::Element;
-use ElementExt;
+use elementext::ElementExt;
 
 use types::{Parse, Encode};
 use error::*;
@@ -68,7 +68,7 @@ impl Encode for ModifiedWriteValues {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use types::test;
+    use run_test;
 
     #[test]
     fn decode_encode() {
@@ -80,6 +80,6 @@ mod tests {
             ),
         ];
         
-        test::<ModifiedWriteValues>(&tests[..]);
+        run_test::<ModifiedWriteValues>(&tests[..]);
     }
 }
