+++ src/svd/register.rs
@@ -1,7 +1,7 @@
 use std::ops::Deref;
 
 use xmltree::Element;
-use ElementExt;
+use elementext::ElementExt;
 
 use types::{Parse, Encode};
 use ::error::{SVDError};
